+++ CHANGELOG.md
@@ -21,7 +21,9 @@ Semantic versioning in our case means:
 
 ### Features
 
+- Adds `python3.15` official support
 - Adds `WPS482`: forbid lazy imports, #3639
+- Adds `WPS483`: forbid `MappingProxyType` on `python3.15+`, #3762
 
 ### Misc